# entropy: fix race condition, massive performance boost