--- a/entropy/entropy_test.go
+++ b/entropy/entropy_test.go
@@ -8,7 +8,8 @@ import (
 
 var dupCount = 0
 
-func check[T comparable](zero T, one T, t *testing.T) {
+func check[T comparable](t *testing.T, zero T, one T) {
+	t.Helper()
 	if zero == one {
 		dupCount++
 		t.Errorf("hit a duplicate! %v == %v", zero, one)
@@ -17,11 +18,14 @@ func check[T comparable](zero T, one T, t *testing.T) {
 }
 
 func Test_RNG(t *testing.T) {
+	t.Parallel()
 	// for coverage
-	sharedRand = GetSharedRand()
+	setSharedRand()
 	RandSleepMS(5)
+	hardLocc.Lock()
 	sharedRand = nil
 	getSharedRand = &sync.Once{}
+	hardLocc.Unlock()
 	//  - - - - - -
 	if OneInA(1000000) {
 		println(string([]byte{
@@ -31,13 +35,14 @@ func Test_RNG(t *testing.T) {
 		}))
 	}
 
-	for n := 0; n != 500; n++ {
-		check(RNG(55555), RNG(55555), t)
-		check(RNGUint32(), RNGUint32(), t)
+	for n := 0; n != 55555; n++ {
+		check(t, RNG(123454321), RNG(123454321))
+		check(t, RNGUint32(), RNGUint32())
 	}
 }
 
 func Test_OneInA(t *testing.T) {
+	t.Parallel()
 	for n := 0; n < 100; n++ {
 		yes := ""
 		if OneInA(1) {
@@ -49,42 +54,48 @@ func Test_OneInA(t *testing.T) {
 	}
 }
 
-func randStrChecks(zero, one string, t *testing.T, intendedLength int) {
+func randStrChecks(t *testing.T, zero, one string, intendedLength int) {
+	t.Helper()
 	if len(zero) != len(one) {
 		t.Fatalf("RandStr output length inconsistency, len(zero) is %d but wanted len(one) which is %d", len(zero), len(one))
 	}
 	if len(zero) != intendedLength || len(one) != intendedLength {
-		t.Fatalf("RandStr output length inconsistency, len(zero) is %d and len(one) is %d, but both should have been 55", len(zero), len(one))
+		t.Fatalf(
+			"RandStr output length inconsistency, "+
+				"len(zero) is %d and len(one) is %d, but both should have been 55", len(zero), len(one))
 	}
-	check(zero, one, t)
+	check(t, zero, one)
 }
 
 func Test_RandStr(t *testing.T) {
+	t.Parallel()
 	for n := 0; n != 500; n++ {
 		zero := RandStr(55)
 		one := RandStr(55)
 		t.Logf("Random0: %s Random1: %s", zero, one)
-		randStrChecks(zero, one, t, 55)
+		randStrChecks(t, zero, one, 55)
 	}
 	t.Logf("[SUCCESS] RandStr had no collisions")
 }
 
 func Test_RandStrWithUpper(t *testing.T) {
+	t.Parallel()
 	for n := 0; n != 500; n++ {
 		zero := RandStrWithUpper(15)
 		one := RandStrWithUpper(15)
 		t.Logf("Random0: %s Random1: %s", zero, one)
-		randStrChecks(zero, one, t, 15)
+		randStrChecks(t, zero, one, 15)
 	}
 	t.Logf("[SUCCESS] RandStr had no collisions")
 }
 
 func Test_RandStr_Entropy(t *testing.T) {
+	t.Parallel()
 	var totalScore = 0
 	for n := 0; n != 500; n++ {
 		zero := RandStr(55)
 		one := RandStr(55)
-		randStrChecks(zero, one, t, 55)
+		randStrChecks(t, zero, one, 55)
 		zeroSplit := strings.Split(zero, "")
 		oneSplit := strings.Split(one, "")
 		var similarity = 0
@@ -93,10 +104,10 @@ func Test_RandStr_Entropy(t *testing.T) {
 				continue
 			}
 			similarity++
-			// t.Logf("[-] zeroSplit[%d] is the same as oneSplit[%d] (%s)", i, i, char)
 		}
 		if similarity*4 > 55 {
-			t.Errorf("[ENTROPY FAILURE] more than a quarter of the string is the same!\n zero: %s \n one: %s \nTotal similar: %d",
+			t.Errorf("[ENTROPY FAILURE] more than a quarter of the string is the same!\n "+
+				"zero: %s \n one: %s \nTotal similar: %d",
 				zero, one, similarity)
 		}
 		// t.Logf("[ENTROPY] Similarity score (lower is better): %d", similarity)
@@ -106,6 +117,7 @@ func Test_RandStr_Entropy(t *testing.T) {
 }
 
 func Test_RandomStrChoice(t *testing.T) {
+	t.Parallel()
 	if RandomStrChoice([]string{}) != "" {
 		t.Fatalf("RandomStrChoice returned a value when given an empty slice")
 	}
@@ -113,12 +125,13 @@ func Test_RandomStrChoice(t *testing.T) {
 	for n := 0; n != 500; n++ {
 		slice = append(slice, RandStr(555))
 	}
-	check(RandomStrChoice(slice), RandomStrChoice(slice), t)
+	check(t, RandomStrChoice(slice), RandomStrChoice(slice))
 }
 
 func Test_RNGUint32(t *testing.T) {
+	t.Parallel()
 	// start globals fresh, just for coverage.
-	sharedRand = GetOptimizedRand()
+	setSharedRand()
 	getSharedRand = &sync.Once{}
 	RNGUint32()
 }